requirements.txt
beautifulsoup4==4.11.1
requests==2.28.0
Django==4.0.5
djangorestframework==3.13.1
PyMySQL==1.0.2
cryptography==37.0.2
redis==4.3.4
celery==5.2.7
pyjwt==2.4.0
python-dotenv==0.20.0